renamed script action package
