+++ src/ScriptActions/FlagStrategy.java
@@ -1,4 +1,4 @@
-package Scripting;
+package ScriptActions;
 
 public enum FlagStrategy {
     AND, OR
